+++ History.md
@@ -5,6 +5,7 @@
 * Breaks 2.x compatibility for using CleanCSS as a function.
 * Reworks minification to tokenize first then minify.
   See [changes](https://github.com/jakubpawlowicz/clean-css/compare/b06f37d...dd8c14a).
+* Renames `noAdvanced` option into `advanced`.
 * Speeds up advanced processing by shortening optimize loop.
 * Fixed issue [#360](https://github.com/GoalSmashers/clean-css/issues/360) - adds 7 extra CSS colors.